move create_shape_aligned_climatology to utils
