+++ src/utils.py
@@ -7,6 +7,7 @@ from collections import namedtuple
 import calendar
 from datetime import date
 import xarray as xr
+import numpy as np
 
 from typing import Optional, Tuple
 
@@ -68,6 +69,77 @@ def get_ds_mask(ds: xr.Dataset) -> xr.Dataset:
     mask.name = 'mask'
 
     return mask
+
+
+def create_shape_aligned_climatology(ds, clim, variable, time_period):
+    """match the time dimension of `clim` to the shape of `ds` so that can
+    perform simple calculations / arithmetic on the values of clim
+
+    Arguments:
+    ---------
+    ds : xr.Dataset
+        the dataset with the raw values that you are comparing to climatology
+
+    clim: xr.Dataset
+        the climatology values for a given `time_period`
+
+    variable: str
+        name of the variable that you are comparing to the climatology
+
+    time_period: str
+        the period string used to calculate the climatology
+         time_period = {'dayofyear', 'season', 'month'}
+
+    Notes:
+        1. assumes that `lat` and `lon` are the
+        coord names in ds
+
+    """
+    for coord in ['lat', 'lon']:
+        assert coord in [c for c in ds.coords]
+
+    ds[time_period] = ds[f'time.{time_period}']
+
+    values = clim[variable].values
+    keys = clim[time_period].values
+    # map the `time_period` to the `values` of the climatology (threshold or mean)
+    lookup_dict = dict(zip(keys, values))
+
+    # extract an array of the `time_period` values from the `ds`
+    timevals = ds[time_period].values
+
+    # use the lat lon arrays (climatology values) in `lookup_dict` corresponding
+    #  to time_period values defined in `timevals` and stack into new array
+    new_clim_vals = np.stack([lookup_dict[timevals[i]] for i in range(len(timevals))])
+
+    assert new_clim_vals.shape == ds[variable].shape, f"\
+        Shapes for new_clim_vals and ds must match! \
+         new_clim_vals.shape: {new_clim_vals.shape} \
+         ds.shape: {ds[variable].shape}"
+
+    # copy that forward in time
+    new_clim = xr.Dataset(
+        {variable: (['time', 'lat', 'lon'], new_clim_vals)},
+        coords={
+            'lat': clim.lat,
+            'lon': clim.lon,
+            'time': ds.time,
+        }
+    )
+
+    return new_clim
+
+
+def drop_nans_and_flatten(dataArray: xr.DataArray) -> np.ndarray:
+    """flatten the array and drop nans from that array. Useful for plotting histograms.
+
+    Arguments:
+    ---------
+    : dataArray (xr.DataArray)
+        the DataArray of your value you want to flatten
+    """
+    # drop NaNs and flatten
+    return dataArray.values[~np.isnan(dataArray.values)]
 
 
 # dictionary lookup of regions
